+++ pkg/qrcodegenerator/qrcodegenerator.go
@@ -23,6 +23,7 @@ import (
 	"go.thethings.network/lorawan-stack/v3/pkg/errors"
 	"go.thethings.network/lorawan-stack/v3/pkg/log"
 	"go.thethings.network/lorawan-stack/v3/pkg/qrcodegenerator/qrcode/enddevices"
+	"go.thethings.network/lorawan-stack/v3/pkg/qrcodegenerator/qrcode/gateways"
 	"go.thethings.network/lorawan-stack/v3/pkg/ttnpb"
 	"google.golang.org/grpc"
 )
@@ -35,23 +36,28 @@ type QRCodeGenerator struct {
 	ctx context.Context
 
 	endDevices *enddevices.Server
+	gateways   *gateways.Server
 
 	grpc struct {
 		endDeviceQRCodeGenerator *endDeviceQRCodeGeneratorServer
+		gatewayQRCodeGenerator   *gatewayQRCodeGeneratorServer
 	}
 }
 
 var errFormatNotFound = errors.DefineNotFound("format_not_found", "format `{id}` not found")
 
 // New returns a new *QRCodeGenerator.
-func New(c *component.Component, conf *Config, opts ...Option) (*QRCodeGenerator, error) {
+func New(c *component.Component, _ *Config, opts ...Option) (*QRCodeGenerator, error) {
 	ctx := log.NewContextWithField(c.Context(), "namespace", "qrcodegenerator")
 	qrg := &QRCodeGenerator{
 		Component: c,
 		ctx:       ctx,
 	}
 	qrg.grpc.endDeviceQRCodeGenerator = &endDeviceQRCodeGeneratorServer{QRG: qrg}
+	qrg.grpc.gatewayQRCodeGenerator = &gatewayQRCodeGeneratorServer{QRG: qrg}
+
 	qrg.endDevices = enddevices.New(ctx)
+	qrg.gateways = gateways.New()
 
 	c.RegisterGRPC(qrg)
 
@@ -72,22 +78,31 @@ func WithEndDeviceFormat(id string, f enddevices.Format) Option {
 	}
 }
 
+// WithGatewayFormat configures QRCodeGenerator with a GatewayFormat.
+func WithGatewayFormat(id string, f gateways.Format) Option {
+	return func(qrg *QRCodeGenerator) {
+		qrg.gateways.RegisterGatewayFormat(id, f)
+	}
+}
+
 // Context returns the context of the QR Code Generator.
 func (qrg *QRCodeGenerator) Context() context.Context {
 	return qrg.ctx
 }
 
 // Roles returns the roles that the QR Code Generator fulfills.
-func (qrg *QRCodeGenerator) Roles() []ttnpb.ClusterRole {
+func (*QRCodeGenerator) Roles() []ttnpb.ClusterRole {
 	return []ttnpb.ClusterRole{ttnpb.ClusterRole_QR_CODE_GENERATOR}
 }
 
 // RegisterServices registers services provided by qrg at s.
 func (qrg *QRCodeGenerator) RegisterServices(s *grpc.Server) {
 	ttnpb.RegisterEndDeviceQRCodeGeneratorServer(s, qrg.grpc.endDeviceQRCodeGenerator)
+	ttnpb.RegisterGatewayQRCodeGeneratorServer(s, qrg.grpc.gatewayQRCodeGenerator)
 }
 
 // RegisterHandlers registers gRPC handlers.
 func (qrg *QRCodeGenerator) RegisterHandlers(s *runtime.ServeMux, conn *grpc.ClientConn) {
-	ttnpb.RegisterEndDeviceQRCodeGeneratorHandler(qrg.Context(), s, conn)
+	ttnpb.RegisterEndDeviceQRCodeGeneratorHandler(qrg.Context(), s, conn) //nolint:errcheck
+	ttnpb.RegisterGatewayQRCodeGeneratorHandler(qrg.Context(), s, conn)   //nolint:errcheck
 }